Many files: Fixed types, formatting and small bugs

+++ src/app/api/applies/route.ts
@@ -1,3 +1,4 @@
+// TODO: 
 import { NextRequest, NextResponse } from "next/server";
 import dbConnect from "@/lib/dbConnect";
 import Apply from "@/models/apply";
@@ -5,31 +6,44 @@ import handleError from "@/helpers/handleError";
 import path from "path";
 import fs from "fs/promises";
 
+type ApplySchemaType = {
+  _id: string;
+  deptid: string;
+  designation: string;
+  experience: string;
+  qualification: string;
+  jobdesc: string;
+  name: string;
+  email: string;
+  mobilenumber: number;
+  resume: string;
+};
 
-export async function GET(request: NextRequest) {
-
-
+// export async function GET(request: NextRequest) {
+export async function GET() {
   try {
     await dbConnect();
-    const applies = await Apply.find({ });
+    const applies = await Apply.find<ApplySchemaType>({});
 
     if (!applies) {
-      return NextResponse.json({ error: "responses not found" }, { status: 404 });
+      return NextResponse.json(
+        { error: "responses not found" },
+        { status: 404 },
+      );
     }
 
     return NextResponse.json(applies);
   } catch (error) {
     return handleError(error, "Server error");
-    
   }
 }
 
 export async function DELETE(request: NextRequest) {
+  // TODO: Add a admin check here and handle that in the frontend
   const { searchParams } = new URL(request.url);
   const id = searchParams.get("id");
-  const resume = searchParams.get("resume");
+  const resume: string | null = searchParams.get("resume");
   try {
-
     await dbConnect();
 
     // Delete award from DB
@@ -41,28 +55,36 @@ export async function DELETE(request: NextRequest) {
           success: false,
           message: "response not found.",
         },
-        { status: 404 }
+        { status: 404 },
       );
     }
 
     // Delete image from filesystem
-    const fullPdfPath = path.join(process.cwd(), "public", resume);
+    // const fullPdfPath = path.join(process.cwd(), "public", resume);
+    let fullPdfPath;
+    if (resume) {
+      fullPdfPath = path.join(process.cwd(), "public", resume);
+    }
+    if (!fullPdfPath) {
+      return NextResponse.json({
+        success: false,
+        message: "There was an error in deleting the pdf",
+      });
+    }
     try {
       await fs.unlink(fullPdfPath);
       console.log("pdf deleted:", fullPdfPath);
     } catch (fsErr) {
-  const err = fsErr as Error;
-  console.warn("Failed to delete pdf:", err.message);
-}
+      const err = fsErr as Error;
+      console.warn("Failed to delete pdf:", err.message);
+    }
 
     return NextResponse.json({
       success: true,
       message: "Response deleted successfully.",
       deletedResponse,
     });
-
   } catch (error) {
     return handleError(error, "Failed to delete Response");
   }
 }
-
